Added a new ruleset (ARS with 16 pieces) and a new mode to go with it (TAP Master with revised grading/requirements)
This is a much improved version of the pseudopolymino mode I submitted on discord a few days ago. 2 more pieces are added and there is an improved rotation system. The approach of having a separate randomizer was discarded in favor of exploiting duplicate pieces to produce a 25-bag randomizer. I have extensively tested this, so it should be fit to include in the main modpack.